Remove Context API and Moment.js usage
As I have become more familiar with Recoil.js and how it can be used for state
management, this commit removes the usage of the React Context API. By doing so,
I am able to simplify the state of this application and the JSX used to inherit
the state's information.

Also, this commit removes Moment.js as a dependency in favor of `date-fns` which
is able to provide our application with the same functionality but with some
performance improvements.