Add Settings panel
This commit adds a new Settings panel to the App. From within the Settings
panel, the user is able to toggle and update a range of different settings.

Previously, this was all handled within the UI at the top level of the app. As
we build out more features and more settings, it makes sense to hide them away
in a separate settings panel.