Reduce the amount of data fetch from Slack users.info
This commit adjusts the `fetchUserProfile` function to reduce the amount of data
we actually need to access and pass to the front-end. Previously, the function
just returned the raw response of the endpoint without any translation.

Now, we only return the fields we need to access in a more accessible and easy
to work with object.