Add Chat with user functionality
This commit adds a feature which allows the user to open a Slack Direct Message
with the User they clicked on. To open a message, click on the small Chat bubble
icon next to the users Timezone flag.

This feature uses Slack's deep-linking feature which uses URI's. Documentation
can be found here: https://api.slack.com/reference/deep-linking#client__supported-uris__open-a-direct-message